backend/llm/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
wandb
weave
httpx
aiofiles
